fix: change DataFrame column type casting from uint16 to int32 for improved compatibility

+++ pyblinkers/fit_blink.py
@@ -100,10 +100,10 @@ class FitBlinks:
             self.frame_blinks = pd.DataFrame()  # <-- create empty DataFrame
             return
 
-        # Cast all columns to uint16 except 'maxValue' (which holds float precision signal peak)
+        # Cast all columns to int32 except 'maxValue' (which holds float precision signal peak)
         # since all positions are always positive and don't require full 64-bit precision
         self.df = self.df.astype({
-            col: 'uint16' for col in self.df.columns if col != 'maxValue'
+            col: 'int32' for col in self.df.columns if col != 'maxValue'
         })
         self.fit()
